feat: app code refactored
