+++ README.md
@@ -1,1 +1,19 @@
 Tasty-Bites - A food delivery app
+
+-
+- Header
+- - Logo
+- - Nav Items
+- Body
+- - Search
+- - RestaurantContainer
+-     - RestaurantCard
+-        - img
+-        - name of res
+-        - star rating
+-        - cuisines
+- Footer
+- - Copyright
+- - Links
+- - Address
+- - Contact us
