+++ src/components/Header.jsx
@@ -1,0 +1,20 @@
+import { logo } from "../utils/images";
+const Header = () => {
+  return (
+    <div className="header">
+      <div className="logo-container">
+        <img className="logo" src={logo} alt="logo" />
+      </div>
+      <div className="nav-items-container">
+        <ul className="nav-items">
+          <li>Home</li>
+          <li>About Us</li>
+          <li>Contact Us</li>
+          <li>Cart</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
