+++ src/utils/images.js
@@ -1,0 +1,3 @@
+import logo from "./../../public/images/logo.png";
+
+export { logo };
