+++ src/utils/mockData.js
@@ -1,29 +1,4 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import logo from "./public/images/logo.png";
-import pizzaHut from "./public/images/pizza-hut.png";
-
-/**
- *
- * Header
- *   - Logo
- *   - Nav Items
- * Body
- *   - Search
- *   - RestaurantContainer
- *     - RestaurantCard
- *        - img
- *        - name of res
- *        - star rating
- *        - cuisines
- * Footer
- *   - Copyright
- *   - Links
- *   - Address
- *   - Contact us
- */
-
-const restrauntList = [
+export const restaurantList = [
   {
     type: "restaurant",
     info: {
@@ -1707,69 +1682,3 @@ const restrauntList = [
     bottomContainers: [],
   },
 ];
-
-const Header = () => {
-  return (
-    <div className="header">
-      <div className="logo-container">
-        <img className="logo" src={logo} alt="logo" />
-      </div>
-      <div className="nav-items-container">
-        <ul className="nav-items">
-          <li>Home</li>
-          <li>About Us</li>
-          <li>Contact Us</li>
-          <li>Cart</li>
-        </ul>
-      </div>
-    </div>
-  );
-};
-
-const RestaurantCard = ({ data }) => {
-  return (
-    <div className="res-card">
-      <div className="res-img-container">
-        <img className="res-img" src={data.info.image.url} alt="pizza-hut" />
-      </div>
-      <div className="food-info">
-        <p>
-          <strong>{data.info.name}</strong>
-        </p>
-        <p>{data.info.rating.aggregate_rating}</p>
-      </div>
-      <div className="food-cost">
-        <p>{data.info.cuisine[0].name}</p>
-        <p>{data.info.costText.text}</p>
-      </div>
-      <div className="time-to-deliver">
-        <p>{data.order.deliveryTime}</p>
-      </div>
-    </div>
-  );
-};
-
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="search">Search</div>
-      <div className="res-container">
-        {restrauntList.map((item) => (
-          <RestaurantCard data={item} key={item.info.resId} />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const AppLayout = () => {
-  return (
-    <div className="app">
-      <Header />
-      <Body />
-    </div>
-  );
-};
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<AppLayout />);
